requirements.txt
uuid
pickledb
grpcio-tools

